Avoid overriding gup methods on composite gup
Prior to this change, the return composite gup object from gupCompose is
extended to have the gup methods (pre, enter, etc.) from the provided
gup fns. This results in the composite gup methods becoming set on the
provided gup fns instead of a new gup instance (as intented).

This change skips the gup methods during extend. This effectively
creates a shallow copy of a gup instance if it is the only gup provided
in fns.

Note that the composite is 'live' in the sense that if any of the gup
methods change on any of the provided fns then this change is reflected
in the composite.