Allow GUP methods alter the transition
Prior to this change, the transition that was in effect when the GUP was
applied was carried forward to every phase of the GUP. Although,
individual GUP methods could remove/alter the transition these
transition were not composable.

With this change gup is cabable of handling any transition that might be
produced from select and/or enter. Any transitions encountered will
replace the existing transition. This means that select is able to
change the transition of pre and exit, and enter is able to change the
post transition.