Remove hard dependencies to D3
Prior to this change, d3-selection and, more notably, d3-transition were
direct dependencies of d3-gup; there's actually little need for this to
be the case.

This change moves those dependencies to peer dependencies, and in doing
so removes support of the direct invocation style of d3-gup; this is a
breaking change.

Typical usage involves invocation of `call`, e.g.

    d3.select(el).call(gup()(data))

Prior to this change, users could also invoke gup like this:

    gup()(d3.select(el), data)

Although arguably convenient, still somewhat confusing.

And this still works}:

    gup()(data)(d3.select(el))